refactor(access): Update Model Person to use PyTest API Fields Render
ref: #761 #730
