+++ app/access/tests/unit/person/conftest.py
@@ -1,0 +1,14 @@
+import pytest
+
+from access.models.person import Person
+
+
+
+@pytest.fixture( scope = 'class')
+def model(request):
+
+    request.cls.model = Person
+
+    yield request.cls.model
+
+    del request.cls.model
